feat: Simplify Goals & Projects tool
