+++ assets/js/components/charts/constants.tsx
@@ -1,0 +1,6 @@
+export const COLORS = {
+  green: "rgb(22, 163, 74)",
+  yellow: "rgb(250, 204, 21)",
+  red: "rgb(239, 68, 68)",
+  gray: "rgb(107, 114, 128)",
+};
