+++ assets/js/components/charts/index.tsx
@@ -1,0 +1,3 @@
+export { PieChart } from "./PieChart";
+export { MiniPieChart } from "./MiniPieChart";
+export { ProgressBar } from "./ProgressBar";
