+++ assets/js/features/goals/GoalTree/components/GoalDetails.tsx
@@ -10,8 +10,7 @@ import { IconCalendar, IconMinus, IconPlus } from "@tabler/icons-react";
 import { includesId, Paths } from "@/routes/paths";
 import { createTestId } from "@/utils/testid";
 import { assertPresent } from "@/utils/assertions";
-import { ProgressBar } from "@/components/ProgressBar";
-import { MiniPieChart } from "@/components/MiniPieChart";
+import { ProgressBar, MiniPieChart } from "@/components/charts";
 import { SecondaryButton } from "@/components/Buttons";
 import { DivLink } from "@/components/Link";
 import { AvatarLink } from "@/components/Avatar";
